Port the shared GDoc to this website.
Add checklists for badge criteria, and help for authors and evaluators.

I also added a section about the basics of Docker to the authors help.

As part of this, I removed the redirects to the old SOSP2019 website,
and added a custom 404 page to ensure anyone following those links
doesn't have a terrible experience... I could revert this and instead
put the new pages in a subfolder, if keeping these old links alive is
considered important.